requirements.txt
numpy==1.23.5
pandas==1.5.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
pillow==9.5.0  # Adjusted version to be compatible with all packages
matplotlib==3.7.3
scikit-image==0.20.0
opencv-python==4.8.1.78
librosa==0.10.0.post2
lime==0.2.0.1
streamlit==1.26.0
